refactor main function
