+++ src/lib.rs
@@ -23,34 +23,4 @@ mod domain;
 mod infrastructure;
 mod presentation;
 
-mod souko {
-    use color_eyre::eyre::Result;
-
-    use crate::{application::service::Service, infrastructure, presentation::Presentation};
-
-    #[derive(Debug)]
-    pub struct Souko {
-        presentation: Presentation,
-    }
-
-    impl Souko {
-        pub fn from_env() -> Self {
-            let presentation = Presentation::from_env();
-            Self { presentation }
-        }
-
-        pub fn command() -> clap::Command {
-            Presentation::command()
-        }
-
-        pub fn verbosity(&self) -> Option<tracing::Level> {
-            self.presentation.verbosity()
-        }
-
-        pub fn main(self) -> Result<()> {
-            let repository = infrastructure::repository();
-            let service = Service::new(&repository);
-            self.presentation.main(&service)
-        }
-    }
-}
+mod souko;
